restructured dir structure
