--- adk/manifest/__init__.py
@@ -1,2 +1,0 @@
-from .modeldata import ModelData
-from .classes import FileData
